requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
groq>=0.9.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
langgraph>=0.1.0
langchain-groq>=1.1.2
langchain-core>=1.3.2
